modification de ft_memchr.c

+++ ft_memchr.c
@@ -6,7 +6,7 @@
 /*   By: tarcay <tarcay@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/12/20 11:37:44 by tarcay            #+#    #+#             */
-/*   Updated: 2020/12/30 22:00:18 by tarcay           ###   ########.fr       */
+/*   Updated: 2020/12/30 22:02:38 by tarcay           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 void	*ft_memchr(const void *s, int c, size_t n)
 {
 	char	*str;
-	int		i;
+	size_t	i;
 
 	str = (char *)s;
 	i = 0;
